Refactor voxel to its own subpackage
