+++ src/main/java/com/gorman/voxel_engine/world/terrain/Chunk.java
@@ -3,7 +3,7 @@ package com.gorman.voxel_engine.world.terrain;
 import java.util.ArrayList;
 
 import com.gorman.voxel_engine.world.primitives.Vector;
-import com.gorman.voxel_engine.world.primitives.Voxel;
+import com.gorman.voxel_engine.world.voxels.Voxel;
 
 /**
  * The Chunk object represents a section of the voxel terrain.
